python/requirements.txt
tika
mysql-connector-python
